commit 15/11/2017 23:55

+++ projet_AM_LA - Copie/map.h
@@ -20,7 +20,7 @@
 void ma(char** tab);
 char** initialisation_tableau( int x , int y  );
 void afficher_power(sprite_t *glass, char** tab, SDL_Renderer *renderer );
-void Afficher(SDL_Rect *R_tileset,SDL_Texture* T_tileset,char** table,int nombre_blocs_largeur,int nombre_blocs_hauteur, SDL_Renderer *renderer, int scrolling_x,int *cpt  );
+void Afficher(SDL_Rect *R_tileset,SDL_Texture* T_tileset,char** table,int nombre_blocs_largeur,int nombre_blocs_hauteur, SDL_Renderer *renderer, float scrolling_x,int *cpt  );
 void modifTabPower(mouse m ,char** tab);
 void anim( SDL_Rect *block, int *cpt, int nombre_sprite, int ligne_sprite );
